feat(api): moves models into quest sub folder
