+++ api/src/quest/domain/models/quest/DataForQuest.js
