+++ api/src/quest/domain/usecases/get-combined-course-blueprint-by-id.js
@@ -1,6 +1,6 @@
 import { NotFoundError } from '../../../shared/domain/errors.js';
 import { AdminCombinedCourseBlueprintDetails } from '../models/AdminCombinedCourseBlueprintDetails.js';
-import { REQUIREMENT_TYPES } from '../models/Quest.js';
+import { REQUIREMENT_TYPES } from '../models/quest/Quest.js';
 
 export const getCombinedCourseBlueprintById = async ({
   id,
