+++ api/src/quest/domain/usecases/reward-user.js
@@ -1,4 +1,4 @@
-import { DataForQuest } from '../models/DataForQuest.js';
+import { DataForQuest } from '../models/quest/DataForQuest.js';
 
 export const rewardUser = async ({
   userId,
