+++ api/src/quest/infrastructure/repositories/combined-course-repository.js
@@ -2,7 +2,7 @@ import { DomainTransaction } from '../../../shared/domain/DomainTransaction.js';
 import { NotFoundError } from '../../../shared/domain/errors.js';
 import { fetchPage } from '../../../shared/infrastructure/utils/knex-utils.js';
 import { CombinedCourse } from '../../domain/models/combined-course/CombinedCourse.js';
-import { Quest } from '../../domain/models/Quest.js';
+import { Quest } from '../../domain/models/quest/Quest.js';
 
 const getByCode = async ({ code }) => {
   const knexConn = DomainTransaction.getConnection();
