+++ api/tests/evaluation/acceptance/application/assessments/assessment-controller_test.js
@@ -17,7 +17,7 @@ import {
   CRITERION_COMPARISONS,
   REQUIREMENT_COMPARISONS,
   REQUIREMENT_TYPES,
-} from '../../../../../src/quest/domain/models/Quest.js';
+} from '../../../../../src/quest/domain/models/quest/Quest.js';
 import { Assessment } from '../../../../../src/shared/domain/models/Assessment.js';
 import { FRENCH_FRANCE } from '../../../../../src/shared/domain/services/locale-service.js';
 import { CORRELATION_METADATA } from '../../../../../src/shared/infrastructure/execution-context-manager.js';
