+++ api/tests/profile/integration/scripts/share-attestations-to-eligible-organizations_test.js
@@ -7,7 +7,7 @@ import {
   CRITERION_COMPARISONS,
   REQUIREMENT_COMPARISONS,
   REQUIREMENT_TYPES,
-} from '../../../../src/quest/domain/models/Quest.js';
+} from '../../../../src/quest/domain/models/quest/Quest.js';
 import { expect } from '../../../test-helper.js';
 import { databaseBuilder, knex } from '../../../tooling/databases.js';
