+++ api/tests/quest/acceptance/application/quest-route_test.js
@@ -5,7 +5,7 @@ import {
   CRITERION_COMPARISONS,
   REQUIREMENT_COMPARISONS,
   REQUIREMENT_TYPES,
-} from '../../../../src/quest/domain/models/Quest.js';
+} from '../../../../src/quest/domain/models/quest/Quest.js';
 import { expect } from '../../../test-helper.js';
 import { databaseBuilder, knex } from '../../../tooling/databases.js';
 import { generateAuthenticatedUserRequestHeaders } from '../../../tooling/test-utils/http-server.js';
