+++ api/tests/quest/integration/infrastructure/repositories/combined-course-blueprint-repository_test.js
@@ -5,7 +5,7 @@ import {
   CRITERION_COMPARISONS,
   REQUIREMENT_COMPARISONS,
   REQUIREMENT_TYPES,
-} from '../../../../../src/quest/domain/models/Quest.js';
+} from '../../../../../src/quest/domain/models/quest/Quest.js';
 import { QuestInput } from '../../../../../src/quest/domain/models/QuestInput.js';
 import * as combinedCourseBluePrintRepository from '../../../../../src/quest/infrastructure/repositories/combined-course-blueprint-repository.js';
 import { expect } from '../../../../test-helper.js';
