+++ api/tests/quest/unit/domain/models/CombinedCourseBlueprint_test.js
@@ -2,7 +2,7 @@ import { REWARD_TYPES } from '../../../../../src/quest/domain/constants.js';
 import { CombinedCourse } from '../../../../../src/quest/domain/models/combined-course/CombinedCourse.js';
 import { CombinedCourseBlueprint } from '../../../../../src/quest/domain/models/CombinedCourseBlueprint.js';
 import { CombinedCourseBlueprintForUpdate } from '../../../../../src/quest/domain/models/CombinedCourseBlueprintForUpdate.js';
-import { Quest } from '../../../../../src/quest/domain/models/Quest.js';
+import { Quest } from '../../../../../src/quest/domain/models/quest/Quest.js';
 import { ObjectValidationError } from '../../../../../src/shared/domain/errors.js';
 import { expect } from '../../../../test-helper.js';
 import { catchErrSync } from '../../../../tooling/test-utils/error.js';
