+++ api/tests/quest/unit/domain/models/Criterion_test.js
@@ -1,5 +1,5 @@
-import { Criterion } from '../../../../../src/quest/domain/models/Criterion.js';
-import { COMPARISONS as CRITERION_PROPERTY_COMPARISONS } from '../../../../../src/quest/domain/models/CriterionProperty.js';
+import { Criterion } from '../../../../../src/quest/domain/models/quest/Criterion.js';
+import { COMPARISONS as CRITERION_PROPERTY_COMPARISONS } from '../../../../../src/quest/domain/models/quest/CriterionProperty.js';
 import { expect } from '../../../../test-helper.js';
 
 describe('Quest | Unit | Domain | Models | Criterion ', function () {
