+++ api/tests/quest/unit/domain/models/QuestInput_test.js
@@ -5,7 +5,7 @@ import {
   Quest,
   REQUIREMENT_COMPARISONS,
   REQUIREMENT_TYPES,
-} from '../../../../../src/quest/domain/models/Quest.js';
+} from '../../../../../src/quest/domain/models/quest/Quest.js';
 import { QuestInput } from '../../../../../src/quest/domain/models/QuestInput.js';
 import { EntityValidationError } from '../../../../../src/shared/domain/errors.js';
 import { expect } from '../../../../test-helper.js';
