+++ api/tests/quest/unit/domain/models/Success_test.js
@@ -1,4 +1,4 @@
-import { Success } from '../../../../../src/quest/domain/models/Success.js';
+import { Success } from '../../../../../src/quest/domain/models/quest/Success.js';
 import { KnowledgeElement } from '../../../../../src/shared/domain/models/KnowledgeElement.js';
 import { expect } from '../../../../test-helper.js';
