+++ api/tests/quest/unit/infrastructure/serializers/combined-course-blueprint-serializer_test.js
@@ -1,5 +1,5 @@
 import { CombinedCourseBlueprint } from '../../../../../src/quest/domain/models/CombinedCourseBlueprint.js';
-import { Quest } from '../../../../../src/quest/domain/models/Quest.js';
+import { Quest } from '../../../../../src/quest/domain/models/quest/Quest.js';
 import * as combinedCourseBlueprintSerializer from '../../../../../src/quest/infrastructure/serializers/combined-course-blueprint-serializer.js';
 import { expect } from '../../../../test-helper.js';
