+++ api/tests/tooling/domain-builder/factory/build-combined-course-details.js
@@ -2,10 +2,10 @@ import { Campaign } from '../../../../src/quest/domain/models/Campaign.js';
 import { CombinedCourse } from '../../../../src/quest/domain/models/combined-course/CombinedCourse.js';
 import { CombinedCourseDetails } from '../../../../src/quest/domain/models/combined-course-participation/CombinedCourseDetails.js';
 import { CombinedCourseBlueprint } from '../../../../src/quest/domain/models/CombinedCourseBlueprint.js';
-import { DataForQuest } from '../../../../src/quest/domain/models/DataForQuest.js';
-import { Eligibility } from '../../../../src/quest/domain/models/Eligibility.js';
 import { Module } from '../../../../src/quest/domain/models/Module.js';
-import { Quest } from '../../../../src/quest/domain/models/Quest.js';
+import { DataForQuest } from '../../../../src/quest/domain/models/quest/DataForQuest.js';
+import { Eligibility } from '../../../../src/quest/domain/models/quest/Eligibility.js';
+import { Quest } from '../../../../src/quest/domain/models/quest/Quest.js';
 
 function buildCombinedCourse({ name, code, organizationId, questId, baseSurveyUrl } = {}) {
   return new CombinedCourse({
